+++ src/components/common/DatePicker.js
@@ -1,19 +1,39 @@
 import React from 'react';
+import { DatePicker as AntDatePicker, Button } from 'antd';
 import moment from 'moment';
 import PropTypes from 'prop-types';
-import { DatePicker as AntDatePicker } from 'antd';
+import 'components/common/DatePicker.css';
 
 class DatePicker extends React.Component {
     constructor(props) {
         super(props);
 
+        this.state = {
+            open: false
+        };
+
         this.onChange = this.onChange.bind(this);
+        this.onOpenChange = this.onOpenChange.bind(this);
+        this.setDate = this.setDate.bind(this);
     }
 
     onChange(value) {
         if (this.props.onChange) {
             this.props.onChange(value ? value.toISOString() : null);
         }
+    }
+
+    onOpenChange(status) {
+        this.setState({ open: status });
+
+        if (this.props.onOpenChange) {
+            this.props.onOpenChange(status);
+        }
+    }
+
+    setDate(value) {
+        this.onChange(value);
+        this.onOpenChange(false);
     }
 
     render() {
@@ -27,14 +47,52 @@ class DatePicker extends React.Component {
             <AntDatePicker
                 {...wrappedProps}
                 value={value}
-                onChange={this.onChange} />
+                onChange={this.onChange}
+                open={this.state.open}
+                onOpenChange={this.onOpenChange}
+                renderExtraFooter={() => (
+                    <React.Fragment>
+                        <Button
+                            type="link"
+                            size="small"
+                            className="datepicker-button first"
+                            onClick={() => this.setDate(moment().add(1, 'day'))}>
+                            Tomorrow
+                        </Button>
+                        <span>-</span>
+                        <Button
+                            type="link"
+                            size="small"
+                            className="datepicker-button"
+                            onClick={() => this.setDate(moment().add(3, 'day'))}>
+                            3 days
+                        </Button>
+                        <span>-</span>
+                        <Button
+                            type="link"
+                            size="small"
+                            className="datepicker-button"
+                            onClick={() => this.setDate(moment().add(1, 'week'))}>
+                            1 week
+                        </Button>
+                        <span>-</span>
+                        <Button
+                            type="link"
+                            size="small"
+                            className="datepicker-button last"
+                            onClick={() => this.setDate(moment().add(1, 'month'))}>
+                            1 month
+                        </Button>
+                    </React.Fragment>
+                )} />
         );
     }
 }
 
 DatePicker.propTypes = {
     value: PropTypes.string,
-    onChange: PropTypes.func
+    onChange: PropTypes.func,
+    onOpenChange: PropTypes.func
 };
 
 export default DatePicker;