+++ src/components/notes/common/NoteSelect.js
@@ -9,9 +9,15 @@ const NoteSelect = forwardRef(function NoteSelect(props, ref) {
     const value = noteApi.notes.find(note => note.id === props.value) ? props.value : undefined;
 
     return (
-        <Select ref={ref} allowClear={true} {...props} value={value}>
+        <Select
+            ref={ref}
+            allowClear={true}
+            showSearch={true}
+            filterOption={(input, option) => (option.props.title || '').toLowerCase().includes(input)}
+            {...props}
+            value={value}>
             {noteApi.notes.map(note => (
-                <Select.Option key={note.id} value={note.id}>
+                <Select.Option key={note.id} value={note.id} title={note.title}>
                     <Icon icon="circle" color={note.color} text={note.title} />
                 </Select.Option>
             ))}